+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Option to configure different y-levels for different dimensions
 - Added Fabric support
+- Added unique config options for Soul Campfire
 
 ## [v1.21.11-0.1.1] - 2026-01-28